Canvas UI First Commit
Includes the kinetic library and the extension objects for the
workflows.

Includes general ability to add components, work flows and components
inside workflows. As well as connecting them, with a straight arrow.
These elements can be dragged.

There is a e-habitat file, that can be viewed as a test. Note this does
not render with the correct x,y coordinates, there also needs more work
done on the presentation and multiline text.